requirements.txt
# Web interface
streamlit

# Data processing
pandas
openpyxl

# LangChain ecosystem
langchain
langchain-google-genai
langchain-community
langchain-experimental
langchain-text-splitters
langgraph

# Vector database and embeddings
chromadb
sentence-transformers

# PDF processing
pypdf

# Monitoring and tracing
langfuse
# Environment management
python-dotenv

# Utilities
tiktoken
